+++ kratos/utilities/sub_model_parts_list_utility.cpp
@@ -18,7 +18,7 @@
 // External includes 
 
 // Project includes
-#include "utilities/model_part_colors.h"
+#include "utilities/sub_model_parts_list_utility.h"
 
 
 namespace Kratos
